feat 增加获取科室和获取用户列表接口

+++ src/main/java/com/zkyzn/project_manager/services/UserInfoService.java
@@ -14,7 +14,7 @@ public class UserInfoService extends MPJBaseServiceImpl<UserInfoDao, UserInfo> {
      * @param userId 用户Id
      * @return 用户信息
      */
-    public UserInfo GetByUserId(Integer userId){
+    public UserInfo GetByUserId(Long userId){
         return baseMapper.selectById(userId);
     }
